+++ src/main/java/com/hyyh/festa/dto/BoothRankingGetResponse.java
@@ -1,6 +1,6 @@
 package com.hyyh.festa.dto;
 
-import com.hyyh.festa.domain.booth.Booth;
+import com.hyyh.festa.domain.Booth;
 
 public record BoothRankingGetResponse(
